add endpoints services
