--- src/Axios Request/Axios.ts
@@ -1,7 +1,0 @@
-import axios from 'axios';
-
-const CoursesApi = axios.create({
-	baseURL: 'http://localhost:4000',
-});
-
-export default CoursesApi;
