+++ src/types/services/servicesTypes.ts
@@ -1,0 +1,12 @@
+export interface logindata {
+	email: string;
+	password: string;
+}
+export interface apiResponse {
+	successful: string;
+	result: string;
+	user: {
+		name: string;
+		email: string;
+	};
+}
